refactor: change the folder name of the about us page
